requirements.txt
crewai
crewai-tools
python-dotenv
openai
anthropic
openlit